+++ dev/functional/index_sequence_util.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <type_traits>  //  std::index_sequence, std::make_index_sequence
+#include <utility>  //  std::index_sequence, std::make_index_sequence
 
 #include "../functional/cxx_universal.h"
 #ifdef SQLITE_ORM_RELAXED_CONSTEXPR_SUPPORTED
@@ -47,5 +48,22 @@ namespace sqlite_orm {
             return reorder_index_sequence(std::index_sequence<Idx...>{}, std::make_index_sequence<sizeof...(Idx)>{});
         }
 #endif
+
+        template<class... Seq>
+        struct flatten_idxseq {
+            using type = std::index_sequence<>;
+        };
+
+        template<size_t... Ix>
+        struct flatten_idxseq<std::index_sequence<Ix...>> {
+            using type = std::index_sequence<Ix...>;
+        };
+
+        template<size_t... As, size_t... Bs, class... Seq>
+        struct flatten_idxseq<std::index_sequence<As...>, std::index_sequence<Bs...>, Seq...>
+            : flatten_idxseq<std::index_sequence<As..., Bs...>, Seq...> {};
+
+        template<class... Seq>
+        using flatten_idxseq_t = typename flatten_idxseq<Seq...>::type;
     }
 }
