+++ dev/table.h
@@ -11,12 +11,12 @@
 #include "functional/cxx_functional_polyfill.h"
 #include "functional/static_magic.h"
 #include "functional/mpl.h"
-#include "typed_comparator.h"
-#include "tuple_helper/index_sequence_util.h"
+#include "functional/index_sequence_util.h"
 #include "tuple_helper/tuple_filter.h"
 #include "tuple_helper/tuple_traits.h"
 #include "tuple_helper/tuple_iteration.h"
 #include "member_traits/member_traits.h"
+#include "typed_comparator.h"
 #include "type_traits.h"
 #include "constraints.h"
 #include "table_info.h"
